Refactor to move all party actions and reducers to its own folder
