+++ client/src/store/actions/actions.js
@@ -1,7 +1,2 @@
-export const CHARACTER_ADD = 'CHARACTER_ADD';
-export const CHARACTER_REMOVE = 'CHARACTER_REMOVE';
-export const CHARACTER_NAME_UPDATE = 'CHARACTER_NAME_UPDATE';
-export const CHARACTER_SLEEP_TIME_INCREASE = 'CHARACTER_SLEEP_TIME_INCREASE';
-
 export const ADD_NIGHT_WATCH_PLAN = 'ADD_NIGHT_WATCH_PLAN';
 export const FETCH_NIGHT_WATCH_PLAN_FAILED = 'FETCH_NIGHT_WATCH_PLAN_FAILED';
