+++ client/src/store/actions/party/actions.js
@@ -1,0 +1,4 @@
+export const CHARACTER_ADD = 'CHARACTER_ADD';
+export const CHARACTER_REMOVE = 'CHARACTER_REMOVE';
+export const CHARACTER_NAME_UPDATE = 'CHARACTER_NAME_UPDATE';
+export const CHARACTER_SLEEP_TIME_INCREASE = 'CHARACTER_SLEEP_TIME_INCREASE';
